feat: moving window size for PER + game injection possibility
-----moving window size-----
It is now possible to select a window step which will increase the replay memory size towards a max value in steps. This can be helpful as a large replay memory might be needed in later stages while the value and policy targets in early experiences become outdated quickly.

----inject games-----
It is now possible to inject games played by different networks/humans in the replay memory of a new/loaded network. This can be used to aid the network in building a model of the environment which is hard to explore.

A game from OpenAI gym can now be played by you in play_game.py and the experiences saved on disk and then later injected into the training process.